client/pyproject.toml
[project]
name = "user-grpc-client"
version = "0.1.0"
description = "FastAPI client for User gRPC service"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "grpcio>=1.68.1", 
    "grpcio-tools>=1.68.1",
    "uvicorn[standard]>=0.32.1",
    "pydantic>=2.10.4",
    "pydantic[email]>=2.10.4",
    "protobuf>=5.28.3",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.3.4",
    "httpx>=0.28.1",
    "pytest-asyncio>=0.25.0",
    "pyright>=1.1.403",
]
